# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: YolNext API
  description: Comprehensive logistics and transportation management platform API
  version: 1.0.0
  contact:
    name: YolNext Support
    email: support@yolnext.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.yolnext.com
    description: Production server
security:
  - bearerAuth: []
paths:
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - userType
              properties:
                name:
                  type: string
                  minLength: 2
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
                userType:
                  type: string
                  enum: [individual, corporate, nakliyeci, tasiyici]
                  example: individual
                phone:
                  type: string
                  example: "+905551234567"
                company:
                  type: string
                  example: ABC Transport
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User registered successfully
                  user:
                    $ref: '#/components/schemas/User'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: user_123
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        phone:
          type: string
          example: "+905551234567"
        userType:
          type: string
          enum: [individual, corporate, nakliyeci, tasiyici, admin]
          example: individual
        company:
          type: string
          example: ABC Transport
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T10:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-01T10:00:00Z
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        errors:
          type: array
          items:
            type: string
          example: [field1 error, field2 error]
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            example:
              success: false
              message: Authentication required
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            example:
              success: false
              message: Resource already exists
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
            example:
              success: false
              message: Validation failed
              errors: [email is required, password must be at least 6 characters]
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
...